build: Add a little checking to Dosemu build script
1/ Check we are in the correct place to execute the build script

2/ Check the DOS batchfile exists where we expect it to be, before
   launching Dosemu.

+++ SOURCES/src/build.sh
@@ -5,6 +5,15 @@ if [ -z "$REV" ] || [ "$REV" -lt 3684 ]; then
     exit 1
 fi
 
+# ensure we are in right place
+cd "$(dirname "$0")"
+
+# ensure the batchfile is present
+if [ ! -f ./MAKEMOS.BAT ] ; then
+    echo "The DOS batchfile MAKEMOS.BAT was not found in $(pwd)"
+    exit 2
+fi
+
 # need to remove uppercase *.SYS files as dosemu will create the
 # lowercase ones, they may collide
 rm -f kernel/*.SYS
